Control tools done

+++ Services/Models/Applicant.cs
@@ -6,19 +6,26 @@ namespace XebecPortal.UI.Services.Models
     {
         public object Avatar { get; set; }
 
-        [JsonProperty("id")] public long Id { get; set; }
+        [JsonProperty("id")] 
+        public int Id { get; set; }
 
-        [JsonProperty("firstName")] public string FirstName { get; set; }
+        [JsonProperty("firstName")] 
+        public string FirstName { get; set; }
 
-        [JsonProperty("lastName")] public string LastName { get; set; }
+        [JsonProperty("lastName")] 
+        public string LastName { get; set; }
 
-        [JsonProperty("cstMark")] public long CstMark { get; set; }
+        [JsonProperty("cstMark")] 
+        public int CstMark { get; set; }
 
-        [JsonProperty("cstComment")] public string CstComment { get; set; }
+        [JsonProperty("cstComment")] 
+        public string CstComment { get; set; }
 
-        [JsonProperty("interviewRating")] public long InterviewRating { get; set; }
+        [JsonProperty("interviewRating")] 
+        public int InterviewRating { get; set; }
 
-        [JsonProperty("interviewComment")] public string InterviewComment { get; set; }
+        [JsonProperty("interviewComment")] 
+        public string InterviewComment { get; set; }
 
         [JsonProperty("phase")] public string Phase { get; set; }
     }
